Implement this example using Handler instead of deprecated AsyncTask
Some changes were made to MainActivity to ensure that UI modification calls are made on the UI thread.

Additional http requests can be implemented by adding more msg.what to the NetworkHandler
